BackEnd/requirements.txt
Flask
flask-session
flask-cors
pymongo
python-dotenv
requests
google-auth
celery
pillow
